+++ R/utils.R
@@ -44,3 +44,5 @@
   }
   request_date
 }
+
+utils::globalVariables(c("id_row"))